identicon: rename the module
