+++ packages/asset-service/src/service/GenerateAssetIcon.ts
